+++ core/chains/evm/txmgr/broadcaster_test.go
@@ -31,6 +31,7 @@ import (
 	evmconfig "github.com/smartcontractkit/chainlink/v2/core/chains/evm/config"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/gas"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr"
+	txmgrcommon "github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr/txmgr"
 	evmtypes "github.com/smartcontractkit/chainlink/v2/core/chains/evm/types"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest/heavyweight"
@@ -170,7 +171,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -193,7 +194,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			BroadcastAt:        &timeNow,
 			InitialBroadcastAt: &timeNow,
 			Error:              null.String{},
-			State:              txmgr.EthTxUnconfirmed,
+			State:              txmgrcommon.EthTxUnconfirmed,
 		}
 		etxWithError := txmgr.EvmTx{
 			Sequence:       nil,
@@ -203,7 +204,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			Value:          value,
 			FeeLimit:       gasLimit,
 			Error:          null.StringFrom(errStr),
-			State:          txmgr.EthTxFatalError,
+			State:          txmgrcommon.EthTxFatalError,
 		}
 
 		require.NoError(t, txStore.InsertEthTx(&etxUnconfirmed))
@@ -223,7 +224,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			Value:          big.Int(assets.NewEthValue(242)),
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(0, 0),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		ethClient.On("SendTransactionReturnCode", mock.Anything, mock.MatchedBy(func(tx *gethTypes.Transaction) bool {
 			return tx.Nonce() == uint64(2) && tx.Value().Cmp(big.NewInt(242)) == 0
@@ -241,7 +242,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			Value:          value,
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(0, 1),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 			Meta:           &meta,
 		}
 		ethClient.On("SendTransactionReturnCode", mock.Anything, mock.MatchedBy(func(tx *gethTypes.Transaction) bool {
@@ -265,7 +266,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			Value:          value,
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(1, 0),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		ethClient.On("SendTransactionReturnCode", mock.Anything, mock.MatchedBy(func(tx *gethTypes.Transaction) bool {
 			if tx.Nonce() != uint64(1) {
@@ -364,7 +365,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			Value:          big.Int(assets.NewEthValue(142)),
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(0, 0),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		eipTxWithAl := txmgr.EvmTx{
 			FromAddress:          fromAddress,
@@ -373,7 +374,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			Value:                big.Int(assets.NewEthValue(242)),
 			FeeLimit:             gasLimit,
 			CreatedAt:            time.Unix(0, 1),
-			State:                txmgr.EthTxUnstarted,
+			State:                txmgrcommon.EthTxUnstarted,
 			AdditionalParameters: txmgr.EvmAccessListFrom(gethTypes.AccessList{gethTypes.AccessTuple{Address: testutils.NewAddress(), StorageKeys: []gethCommon.Hash{utils.NewHash()}}}),
 		}
 		ethClient.On("SendTransactionReturnCode", mock.Anything, mock.MatchedBy(func(tx *gethTypes.Transaction) bool {
@@ -430,7 +431,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 				Value:          big.Int(assets.NewEthValue(442)),
 				FeeLimit:       gasLimit,
 				CreatedAt:      time.Unix(0, 0),
-				State:          txmgr.EthTxUnstarted,
+				State:          txmgrcommon.EthTxUnstarted,
 				TransmitChecker: checkerToJson(t, txmgr.EvmTransmitCheckerSpec{
 					CheckerType: txmgr.TransmitCheckerTypeSimulate,
 				}),
@@ -464,7 +465,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 			// Check ethtx was sent
 			ethTx, err := txStore.FindEthTxWithAttempts(ethTx.ID)
 			require.NoError(t, err)
-			assert.Equal(t, txmgr.EthTxUnconfirmed, ethTx.State)
+			assert.Equal(t, txmgrcommon.EthTxUnconfirmed, ethTx.State)
 		})
 
 		t.Run("with unknown error, sends tx as normal", func(t *testing.T) {
@@ -475,7 +476,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 				Value:          big.Int(assets.NewEthValue(542)),
 				FeeLimit:       gasLimit,
 				CreatedAt:      time.Unix(0, 0),
-				State:          txmgr.EthTxUnstarted,
+				State:          txmgrcommon.EthTxUnstarted,
 				TransmitChecker: checkerToJson(t, txmgr.EvmTransmitCheckerSpec{
 					CheckerType: txmgr.TransmitCheckerTypeSimulate,
 				}),
@@ -497,7 +498,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 
 			ethTx, err := txStore.FindEthTxWithAttempts(ethTx.ID)
 			require.NoError(t, err)
-			assert.Equal(t, txmgr.EthTxUnconfirmed, ethTx.State)
+			assert.Equal(t, txmgrcommon.EthTxUnconfirmed, ethTx.State)
 		})
 
 		t.Run("on revert, marks tx as fatally errored and does not send", func(t *testing.T) {
@@ -508,7 +509,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 				Value:          big.Int(assets.NewEthValue(642)),
 				FeeLimit:       gasLimit,
 				CreatedAt:      time.Unix(0, 0),
-				State:          txmgr.EthTxUnstarted,
+				State:          txmgrcommon.EthTxUnstarted,
 				TransmitChecker: checkerToJson(t, txmgr.EvmTransmitCheckerSpec{
 					CheckerType: txmgr.TransmitCheckerTypeSimulate,
 				}),
@@ -533,7 +534,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success(t *testing.T) {
 
 			ethTx, err := txStore.FindEthTxWithAttempts(ethTx.ID)
 			require.NoError(t, err)
-			assert.Equal(t, txmgr.EthTxFatalError, ethTx.State)
+			assert.Equal(t, txmgrcommon.EthTxFatalError, ethTx.State)
 			assert.True(t, ethTx.Error.Valid)
 			assert.Equal(t, "transaction reverted during simulation: json-rpc error { Code = 42, Message = 'oh no, it reverted', Data = 'KqYi' }", ethTx.Error.String)
 		})
@@ -568,7 +569,7 @@ func TestEthBroadcaster_TransmitChecking(t *testing.T) {
 			Value:          big.Int(assets.NewEthValue(442)),
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(0, 0),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 			TransmitChecker: checkerToJson(t, txmgr.EvmTransmitCheckerSpec{
 				CheckerType: txmgr.TransmitCheckerTypeSimulate,
 			}),
@@ -587,7 +588,7 @@ func TestEthBroadcaster_TransmitChecking(t *testing.T) {
 		// Check ethtx was sent
 		ethTx, err := txStore.FindEthTxWithAttempts(ethTx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxUnconfirmed, ethTx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnconfirmed, ethTx.State)
 	})
 
 	t.Run("when transmit checking succeeds, sends tx as normal", func(t *testing.T) {
@@ -601,7 +602,7 @@ func TestEthBroadcaster_TransmitChecking(t *testing.T) {
 			Value:          big.Int(assets.NewEthValue(442)),
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(0, 0),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 			TransmitChecker: checkerToJson(t, txmgr.EvmTransmitCheckerSpec{
 				CheckerType: txmgr.TransmitCheckerTypeSimulate,
 			}),
@@ -620,7 +621,7 @@ func TestEthBroadcaster_TransmitChecking(t *testing.T) {
 		// Check ethtx was sent
 		ethTx, err := txStore.FindEthTxWithAttempts(ethTx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxUnconfirmed, ethTx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnconfirmed, ethTx.State)
 	})
 
 	t.Run("when transmit errors, fatally error transaction", func(t *testing.T) {
@@ -634,7 +635,7 @@ func TestEthBroadcaster_TransmitChecking(t *testing.T) {
 			Value:          big.Int(assets.NewEthValue(442)),
 			FeeLimit:       gasLimit,
 			CreatedAt:      time.Unix(0, 0),
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 			TransmitChecker: checkerToJson(t, txmgr.EvmTransmitCheckerSpec{
 				CheckerType: txmgr.TransmitCheckerTypeSimulate,
 			}),
@@ -650,7 +651,7 @@ func TestEthBroadcaster_TransmitChecking(t *testing.T) {
 		// Check ethtx was sent
 		ethTx, err := txStore.FindEthTxWithAttempts(ethTx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxFatalError, ethTx.State)
+		assert.Equal(t, txmgrcommon.EthTxFatalError, ethTx.State)
 		assert.True(t, ethTx.Error.Valid)
 		assert.Equal(t, "fatal checker error", ethTx.Error.String)
 	})
@@ -695,7 +696,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_OptimisticLockingOnEthTx(t *testi
 		EncodedPayload: []byte{42, 42, 0},
 		Value:          big.Int(*assets.NewEth(0)),
 		FeeLimit:       500000,
-		State:          txmgr.EthTxUnstarted,
+		State:          txmgrcommon.EthTxUnstarted,
 	}
 	require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -751,7 +752,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Success_WithMultiplier(t *testing
 		Value:          big.Int(assets.NewEthValue(242)),
 		FeeLimit:       1231,
 		CreatedAt:      time.Unix(0, 0),
-		State:          txmgr.EthTxUnstarted,
+		State:          txmgrcommon.EthTxUnstarted,
 	}
 	require.NoError(t, txStore.InsertEthTx(&tx))
 
@@ -789,7 +790,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_ResumingFromCrash(t *testing.T) {
 			Value:          value,
 			FeeLimit:       gasLimit,
 			Error:          null.String{},
-			State:          txmgr.EthTxInProgress,
+			State:          txmgrcommon.EthTxInProgress,
 		}
 
 		secondInProgress := txmgr.EvmTx{
@@ -800,7 +801,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_ResumingFromCrash(t *testing.T) {
 			Value:          value,
 			FeeLimit:       gasLimit,
 			Error:          null.String{},
-			State:          txmgr.EthTxInProgress,
+			State:          txmgrcommon.EthTxInProgress,
 		}
 
 		require.NoError(t, txStore.InsertEthTx(&firstInProgress))
@@ -1097,7 +1098,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1146,7 +1147,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 				EncodedPayload: encodedPayload,
 				Value:          value,
 				FeeLimit:       gasLimit,
-				State:          txmgr.EthTxUnstarted,
+				State:          txmgrcommon.EthTxUnstarted,
 			}
 			require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1190,7 +1191,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 				EncodedPayload:    encodedPayload,
 				Value:             value,
 				FeeLimit:          gasLimit,
-				State:             txmgr.EthTxUnstarted,
+				State:             txmgrcommon.EthTxUnstarted,
 				PipelineTaskRunID: uuid.NullUUID{UUID: tr.ID, Valid: true},
 			}
 
@@ -1266,7 +1267,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1339,7 +1340,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1363,7 +1364,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.Nil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxInProgress, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxInProgress, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 		attempt := etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptInProgress, attempt.State)
@@ -1387,7 +1388,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.NotNil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxUnconfirmed, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnconfirmed, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 		attempt = etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptBroadcast, attempt.State)
@@ -1403,7 +1404,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1427,7 +1428,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.Nil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxInProgress, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxInProgress, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 		attempt := etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptInProgress, attempt.State)
@@ -1451,7 +1452,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.NotNil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxUnconfirmed, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnconfirmed, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 		attempt = etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptBroadcast, attempt.State)
@@ -1467,7 +1468,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1490,7 +1491,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.NotNil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxUnconfirmed, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnconfirmed, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 		attempt := etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptBroadcast, attempt.State)
@@ -1509,7 +1510,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1554,7 +1555,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		EncodedPayload: encodedPayload,
 		Value:          value,
 		FeeLimit:       gasLimit,
-		State:          txmgr.EthTxUnstarted,
+		State:          txmgrcommon.EthTxUnstarted,
 	}
 	require.NoError(t, txStore.InsertEthTx(&etxUnfinished))
 
@@ -1580,7 +1581,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.Nil(t, etx.InitialBroadcastAt)
 		assert.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxInProgress, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxInProgress, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 		assert.Equal(t, txmgrtypes.TxAttemptInProgress, etx.TxAttempts[0].State)
 	})
@@ -1639,7 +1640,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1667,7 +1668,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1688,7 +1689,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.Nil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxInProgress, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxInProgress, etx.State)
 		require.Len(t, etx.TxAttempts, 1)
 		attempt := etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptInProgress, attempt.State)
@@ -1706,7 +1707,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1726,7 +1727,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 		assert.Nil(t, etx.InitialBroadcastAt)
 		require.NotNil(t, etx.Sequence)
 		assert.False(t, etx.Error.Valid)
-		assert.Equal(t, txmgr.EthTxInProgress, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxInProgress, etx.State)
 		require.Len(t, etx.TxAttempts, 1)
 		attempt := etx.TxAttempts[0]
 		assert.Equal(t, txmgrtypes.TxAttemptInProgress, attempt.State)
@@ -1757,7 +1758,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1789,7 +1790,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_Errors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 		require.NoError(t, txStore.InsertEthTx(&etx))
 
@@ -1871,7 +1872,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_KeystoreErrors(t *testing.T) {
 			EncodedPayload: encodedPayload,
 			Value:          value,
 			FeeLimit:       gasLimit,
-			State:          txmgr.EthTxUnstarted,
+			State:          txmgrcommon.EthTxUnstarted,
 		}
 
 		tx := *gethTypes.NewTx(&gethTypes.LegacyTx{})
@@ -1893,7 +1894,7 @@ func TestEthBroadcaster_ProcessUnstartedEthTxs_KeystoreErrors(t *testing.T) {
 		etx, err = txStore.FindEthTxWithAttempts(etx.ID)
 		require.NoError(t, err)
 
-		assert.Equal(t, txmgr.EthTxUnstarted, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnstarted, etx.State)
 		assert.Len(t, etx.TxAttempts, 0)
 
 		// Check that the key did not have its nonce incremented
