+++ core/chains/evm/txmgr/common.go
@@ -16,9 +16,6 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 )
 
-// timeout value for batchSendTransactions
-const batchSendTransactionTimeout = 30 * time.Second
-
 // Tries to send transactions in batches. Even if some batch(es) fail to get sent, it tries all remaining batches,
 // before returning with error for the latest batch send. If a batch send fails, this sets the error on all
 // elements in that batch.
