+++ core/chains/evm/txmgr/evm_tx_store_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/assets"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/gas"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr"
+	txmgrcommon "github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr/txmgr"
 	evmtypes "github.com/smartcontractkit/chainlink/v2/core/chains/evm/types"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
@@ -52,7 +53,7 @@ func TestORM_EthTransactionsWithAttempts(t *testing.T) {
 
 	// tx 3 has no attempts
 	tx3 := cltest.NewEthTx(t, from)
-	tx3.State = txmgr.EthTxUnstarted
+	tx3.State = txmgrcommon.EthTxUnstarted
 	tx3.FromAddress = from
 	require.NoError(t, orm.InsertEthTx(&tx3))
 
@@ -100,7 +101,7 @@ func TestORM_EthTransactions(t *testing.T) {
 
 	// tx 3 has no attempts
 	tx3 := cltest.NewEthTx(t, from)
-	tx3.State = txmgr.EthTxUnstarted
+	tx3.State = txmgrcommon.EthTxUnstarted
 	tx3.FromAddress = from
 	require.NoError(t, orm.InsertEthTx(&tx3))
 
@@ -217,7 +218,7 @@ func TestORM_FindEthTxAttemptConfirmedByEthTxIDs(t *testing.T) {
 
 	// tx 3 has no attempts
 	tx3 := cltest.NewEthTx(t, from)
-	tx3.State = txmgr.EthTxUnstarted
+	tx3.State = txmgrcommon.EthTxUnstarted
 	tx3.FromAddress = from
 	require.NoError(t, orm.InsertEthTx(&tx3))
 
@@ -364,7 +365,7 @@ func TestORM_UpdateBroadcastAts(t *testing.T) {
 		time1 := time.Now()
 		etx := cltest.NewEthTx(t, fromAddress)
 		etx.Sequence = new(evmtypes.Nonce)
-		etx.State = txmgr.EthTxUnconfirmed
+		etx.State = txmgrcommon.EthTxUnconfirmed
 		etx.BroadcastAt = &time1
 		etx.InitialBroadcastAt = &time1
 		err := orm.InsertEthTx(&etx)
@@ -482,12 +483,12 @@ func TestORM_UpdateEthTxsUnconfirmed(t *testing.T) {
 	originalBroadcastAt := time.Unix(1616509100, 0)
 	etx0 := cltest.MustInsertConfirmedMissingReceiptEthTxWithLegacyAttempt(
 		t, txStore, 0, 1, originalBroadcastAt, fromAddress)
-	assert.Equal(t, etx0.State, txmgr.EthTxConfirmedMissingReceipt)
+	assert.Equal(t, etx0.State, txmgrcommon.EthTxConfirmedMissingReceipt)
 	require.NoError(t, txStore.UpdateEthTxsUnconfirmed([]int64{etx0.ID}))
 
 	etx0, err := txStore.FindEthTxWithAttempts(etx0.ID)
 	require.NoError(t, err)
-	assert.Equal(t, etx0.State, txmgr.EthTxUnconfirmed)
+	assert.Equal(t, etx0.State, txmgrcommon.EthTxUnconfirmed)
 }
 
 func TestORM_FindEthTxAttemptsRequiringReceiptFetch(t *testing.T) {
@@ -542,7 +543,7 @@ func TestORM_SaveFetchedReceipts(t *testing.T) {
 	require.Len(t, etx0.TxAttempts, 1)
 	require.Len(t, etx0.TxAttempts[0].Receipts, 1)
 	require.Equal(t, txmReceipt.BlockHash, etx0.TxAttempts[0].Receipts[0].BlockHash)
-	require.Equal(t, txmgr.EthTxConfirmed, etx0.State)
+	require.Equal(t, txmgrcommon.EthTxConfirmed, etx0.State)
 }
 
 func TestORM_MarkAllConfirmedMissingReceipt(t *testing.T) {
@@ -561,18 +562,18 @@ func TestORM_MarkAllConfirmedMissingReceipt(t *testing.T) {
 	etx0_attempt := newBroadcastLegacyEthTxAttempt(t, etx0.ID, int64(1))
 	etx0_attempt.BroadcastBeforeBlockNum = &etx0_blocknum
 	require.NoError(t, txStore.InsertEthTxAttempt(&etx0_attempt))
-	assert.Equal(t, txmgr.EthTxUnconfirmed, etx0.State)
+	assert.Equal(t, txmgrcommon.EthTxUnconfirmed, etx0.State)
 
 	// create transaction 1 (nonce 1) that is confirmed (block 77)
 	etx1 := cltest.MustInsertConfirmedEthTxBySaveFetchedReceipts(t, txStore, fromAddress, int64(1), int64(77), *ethClient.ConfiguredChainID())
-	assert.Equal(t, etx1.State, txmgr.EthTxConfirmed)
+	assert.Equal(t, etx1.State, txmgrcommon.EthTxConfirmed)
 
 	// mark transaction 0 confirmed_missing_receipt
 	err := txStore.MarkAllConfirmedMissingReceipt(ethClient.ConfiguredChainID())
 	require.NoError(t, err)
 	etx0, err = txStore.FindEthTxWithAttempts(etx0.ID)
 	require.NoError(t, err)
-	assert.Equal(t, txmgr.EthTxConfirmedMissingReceipt, etx0.State)
+	assert.Equal(t, txmgrcommon.EthTxConfirmedMissingReceipt, etx0.State)
 }
 
 func TestORM_PreloadEthTxes(t *testing.T) {
@@ -717,7 +718,7 @@ func TestORM_UpdateEthTxForRebroadcast(t *testing.T) {
 		attempt := etx.TxAttempts[0]
 		require.Equal(t, txmgrtypes.TxAttemptBroadcast, attempt.State)
 		// assert tx state
-		assert.Equal(t, txmgr.EthTxConfirmed, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxConfirmed, etx.State)
 		// assert receipt
 		assert.Len(t, etx.TxAttempts[0].Receipts, 1)
 
@@ -734,7 +735,7 @@ func TestORM_UpdateEthTxForRebroadcast(t *testing.T) {
 		assert.Equal(t, txmgrtypes.TxAttemptInProgress, resultTxAttempt.State)
 		assert.Nil(t, resultTxAttempt.BroadcastBeforeBlockNum)
 		// assert tx state
-		assert.Equal(t, txmgr.EthTxUnconfirmed, resultTx.State)
+		assert.Equal(t, txmgrcommon.EthTxUnconfirmed, resultTx.State)
 		// assert receipt
 		assert.Len(t, resultTxAttempt.Receipts, 0)
 	})
@@ -845,7 +846,7 @@ func TestORM_SaveConfirmedMissingReceiptAttempt(t *testing.T) {
 
 		etx, err := txStore.FindEthTxWithAttempts(etx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxConfirmedMissingReceipt, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxConfirmedMissingReceipt, etx.State)
 		assert.Equal(t, txmgrtypes.TxAttemptBroadcast, etx.TxAttempts[0].State)
 	})
 }
@@ -1034,7 +1035,7 @@ func TestORM_MarkOldTxesMissingReceiptAsErrored(t *testing.T) {
 
 		etx, err = txStore.FindEthTxWithAttempts(etx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxFatalError, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxFatalError, etx.State)
 	})
 
 	t.Run("successfully mark errored transactions w/ qopt passing in sql.Tx", func(t *testing.T) {
@@ -1051,7 +1052,7 @@ func TestORM_MarkOldTxesMissingReceiptAsErrored(t *testing.T) {
 		// must run other query outside of postgres transaction so changes are committed
 		etx, err = txStore.FindEthTxWithAttempts(etx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxFatalError, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxFatalError, etx.State)
 	})
 }
 
@@ -1175,7 +1176,7 @@ func TestORM_UpdateEthTxFatalError(t *testing.T) {
 		etx, err = txStore.FindEthTxWithAttempts(etx.ID)
 		require.NoError(t, err)
 		assert.Len(t, etx.TxAttempts, 0)
-		assert.Equal(t, txmgr.EthTxFatalError, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxFatalError, etx.State)
 	})
 }
 
@@ -1233,7 +1234,7 @@ func TestORM_UpdateEthTxUnstartedToInProgress(t *testing.T) {
 
 		etx, err = txStore.FindEthTxWithAttempts(etx.ID)
 		require.NoError(t, err)
-		assert.Equal(t, txmgr.EthTxInProgress, etx.State)
+		assert.Equal(t, txmgrcommon.EthTxInProgress, etx.State)
 		assert.Len(t, etx.TxAttempts, 1)
 	})
 
@@ -1293,7 +1294,7 @@ func TestORM_UpdateEthTxUnstartedToInProgress(t *testing.T) {
 		etx := cltest.MustInsertInProgressEthTxWithAttempt(t, txStore, nonce, fromAddress)
 		require.Len(t, etx.TxAttempts, 1)
 
-		etx.State = txmgr.EthTxUnstarted
+		etx.State = txmgrcommon.EthTxUnstarted
 
 		// Should fail due to idx_eth_tx_attempt_hash constraint
 		err := txStore.UpdateEthTxUnstartedToInProgress(&etx, &etx.TxAttempts[0])
@@ -1542,7 +1543,7 @@ func TestORM_CreateEthTransaction(t *testing.T) {
 		assert.NoError(t, err)
 
 		assert.Greater(t, etx.ID, int64(0))
-		assert.Equal(t, etx.State, txmgr.EthTxUnstarted)
+		assert.Equal(t, etx.State, txmgrcommon.EthTxUnstarted)
 		assert.Equal(t, gasLimit, etx.FeeLimit)
 		assert.Equal(t, fromAddress, etx.FromAddress)
 		assert.Equal(t, toAddress, etx.ToAddress)
@@ -1555,7 +1556,7 @@ func TestORM_CreateEthTransaction(t *testing.T) {
 		var dbEthTx txmgr.DbEthTx
 		require.NoError(t, db.Get(&dbEthTx, `SELECT * FROM eth_txes ORDER BY id ASC LIMIT 1`))
 
-		assert.Equal(t, dbEthTx.State, txmgr.EthTxUnstarted)
+		assert.Equal(t, dbEthTx.State, txmgrcommon.EthTxUnstarted)
 		assert.Equal(t, gasLimit, dbEthTx.GasLimit)
 		assert.Equal(t, fromAddress, dbEthTx.FromAddress)
 		assert.Equal(t, toAddress, dbEthTx.ToAddress)
